feat(notification): add ActionButton subcomponent
